Created new packages in the ecal-readout-sim project to better separate out classes. There is now org.hps.readout.hodoscope for hodoscope classes, org.hps.readout.trigger for trigger drivers, and org.hps.readout which contains certain base classes that are used for both the hodoscope and calorimeter. The remaining calorimeter classes remain in org.hps.readout.ecal.updated pending the removal of the legacy code in org.hps.readout.ecal.
