+++ ecal-readout-sim/src/main/java/org/hps/readout/SimTrackerHitReadoutDriver.java
@@ -1,35 +1,33 @@
-package org.hps.readout.ecal.updated;
+package org.hps.readout;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.Set;
 
-import org.hps.readout.ReadoutDataManager;
-import org.hps.readout.SLICDataReadoutDriver;
 import org.hps.readout.util.collection.LCIOCollection;
 import org.hps.readout.util.collection.LCIOCollectionFactory;
 import org.hps.readout.util.collection.TriggeredLCIOData;
 import org.lcsim.event.MCParticle;
-import org.lcsim.event.SimCalorimeterHit;
+import org.lcsim.event.SimTrackerHit;
 
 /**
- * <code>SimCalorimeterHitReadoutDriver</code> handles SLIC objects
- * in input Monte Carlo files of type {@link
- * org.lcsim.event.SimCalorimeterHit SimCalorimeterHit}.
+ * <code>SimTrackerHitReadoutDriver</code> handles SLIC objects in
+ * input Monte Carlo files of type {@link
+ * org.lcsim.event.SimTrackerHit SimTrackerHit}.
  * 
  * @author Kyle McCarty <mccarty@jlab.org>
  * @see org.hps.readout.SLICDataReadoutDriver
  */
-public class SimCalorimeterHitReadoutDriver extends SLICDataReadoutDriver<SimCalorimeterHit> {
+public class SimTrackerHitReadoutDriver extends SLICDataReadoutDriver<SimTrackerHit> {
     /**
      * Instantiate an instance of {@link
      * org.hps.readout.SLICDataReadoutDriver SLICDataReadoutDriver}
      * for objects of type {@link
-     * org.lcsim.event.SimCalorimeterHit SimCalorimeterHit} and set
-     * the appropriate LCIO flags.
+     * org.lcsim.event.SimTrackerHit SimTrackerHit
+     * SimCalorimeterHit} and set the appropriate LCIO flags.
      */
-    public SimCalorimeterHitReadoutDriver() {
-        super(SimCalorimeterHit.class, 0xe0000000);
+    public SimTrackerHitReadoutDriver() {
+        super(SimTrackerHit.class, 0xc0000000);
     }
     
     @Override
@@ -39,16 +37,15 @@ public class SimCalorimeterHitReadoutDriver extends SLICDataReadoutDriver<SimCal
         if(!isPersistent()) { return null; }
         
         // Get the truth hits in the indicated time range.
-        Collection<SimCalorimeterHit> truthHits = ReadoutDataManager.getData(triggerTime - getReadoutWindowBefore(), triggerTime + getReadoutWindowAfter(), collectionName, SimCalorimeterHit.class);
+        Collection<SimTrackerHit> truthHits = ReadoutDataManager.getData(triggerTime - getReadoutWindowBefore(), triggerTime + getReadoutWindowAfter(),
+                collectionName, SimTrackerHit.class);
         
         // MC particles need to be extracted from the truth hits
         // and included in the readout data to ensure that the
         // full truth chain is available.
         Set<MCParticle> truthParticles = new java.util.HashSet<MCParticle>();
-        for(SimCalorimeterHit simHit : truthHits) {
-            for(int i = 0; i < simHit.getMCParticleCount(); i++) {
-                ReadoutDataManager.addParticleParents(simHit.getMCParticle(i), truthParticles);
-            }
+        for(SimTrackerHit simHit : truthHits) {
+            ReadoutDataManager.addParticleParents(simHit.getMCParticle(), truthParticles);
         }
         
         // Create the truth MC particle collection.
