+++ ecal-readout-sim/src/main/java/org/hps/readout/trigger/PairTriggerReadoutDriver.java
@@ -1,4 +1,4 @@
-package org.hps.readout.ecal.updated;
+package org.hps.readout.trigger;
 
 import java.util.ArrayList;
 import java.util.Collection;
