+++ ecal-readout-sim/src/main/java/org/hps/readout/trigger/PulserTriggerDriver.java
@@ -1,4 +1,4 @@
-package org.hps.readout.ecal.updated;
+package org.hps.readout.trigger;
 
 import org.hps.readout.TriggerDriver;
 import org.lcsim.event.EventHeader;
